+++ src/skymusic/renderers/instrument_renderers/__init__.py
@@ -1,0 +1,9 @@
+from . import instrument_renderer, html_ir, svg_ir, png_ir, midi_ir, skyjson_ir, ascii_ir
+__all__ = [instrument_renderer, html_ir, svg_ir, png_ir, midi_ir, skyjson_ir, ascii_ir]
+
+import os
+import sys
+
+sys.path.append(os.path.dirname(os.path.abspath(__file__)))
+
+
