+++ src/skymusic/renderers/song_renderers/__init__.py
@@ -1,0 +1,9 @@
+from . import song_renderer, html_sr, svg_sr, png_sr, midi_sr, skyjson_sr, ascii_sr
+__all__ = [song_renderer, html_sr, svg_sr, png_sr, midi_sr, skyjson_sr, ascii_sr]
+
+import os
+import sys
+
+sys.path.append(os.path.dirname(os.path.abspath(__file__)))
+
+
