+++ src/skymusic/renderers/song_renderers/ascii_sr.py
@@ -1,0 +1,46 @@
+import re, io
+from . import song_renderer
+from src.skymusic.renderers.instrument_renderers.ascii_ir import AsciiInstrumentRenderer
+from src.skymusic.resources import Resources
+
+
+class AsciiSongRenderer(song_renderer.SongRenderer):
+
+    def __init__(self, locale=None):
+        
+        super().__init__(locale) 
+        
+    def write_buffers(self, song, render_mode):
+
+        meta = song.get_meta()
+        ascii_buffer = io.StringIO()
+
+        note_parser = render_mode.get_note_parser()
+        instrument_renderer = AsciiInstrumentRenderer(self.locale)
+        
+        ascii_buffer.write(f"{Resources.COMMENT_DELIMITER}{meta['title'][1]}\n")
+
+        for k in meta:
+            if k != 'title':
+                ascii_buffer.write(f"{Resources.COMMENT_DELIMITER}{meta[k][0]}{meta[k][1]}\n")
+
+        
+        song_render = '\n'
+        instrument_index = 0
+        for line in song.get_lines():
+            line_render = ''
+            for instrument in line:
+                instrument.set_index(instrument_index)
+                #instrument_render = instrument.render_in_ascii(note_parser)
+                instrument_render = instrument_renderer.render(instrument, note_parser)                
+                repeat = instrument.get_repeat()
+                if repeat > 1:
+                    instrument_render += Resources.REPEAT_INDICATOR + str(repeat)
+                line_render += instrument_render + re.sub('\\\\s', ' ', Resources.ICON_DELIMITER)
+                instrument_index += 1
+            song_render += '\n' + line_render
+
+        ascii_buffer.write(song_render)
+
+        return [ascii_buffer]
+
